requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
django-admin-numeric-filter==0.1.6
django-jsonfield==1.2.0